# lib/api-zod/src/generated/types/campaignInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CampaignInput {
  /** @minLength 2 */
  productName: string;
  /** @minLength 10 */
  productDescription: string;
  /** @minLength 5 */
  targetAudience: string;
  audienceAge: string;
  audienceLocation: string;
  budget: string;
  /** @nullable */
  customBudget?: string | null;
  campaignGoal: string;
  timeline: string;
  industry: string;
}