# lib/api-zod/src/generated/types/campaignInsight.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CampaignInsightType } from "./campaignInsightType";

export interface CampaignInsight {
  type: CampaignInsightType;
  message: string;
}