# lib/api-zod/src/generated/types/campaignItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CampaignItemContent } from "./campaignItemContent";
import type { CampaignItemEffort } from "./campaignItemEffort";
import type { CampaignItemPriority } from "./campaignItemPriority";
import type { CampaignItemType } from "./campaignItemType";

export interface CampaignItem {
  id: string;
  type: CampaignItemType;
  title: string;
  description: string;
  content: CampaignItemContent;
  priority: CampaignItemPriority;
  effort: CampaignItemEffort;
  score: number;
}