# lib/api-zod/src/generated/types/campaignMetadata.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CampaignMetadataEffort } from "./campaignMetadataEffort";
import type { CampaignMetadataPriority } from "./campaignMetadataPriority";
import type { CampaignMetadataRisk } from "./campaignMetadataRisk";

export interface CampaignMetadata {
  priority: CampaignMetadataPriority;
  risk: CampaignMetadataRisk;
  effort: CampaignMetadataEffort;
  overallScore: number;
  classification: string;
  budgetTier: string;
  goalAlignment: number;
}