# lib/api-zod/src/generated/types/campaignResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CampaignInsight } from "./campaignInsight";
import type { CampaignItem } from "./campaignItem";
import type { CampaignMetadata } from "./campaignMetadata";
import type { ChannelAllocation } from "./channelAllocation";

export interface CampaignResult {
  summary: string;
  items: CampaignItem[];
  insights: CampaignInsight[];
  channelAllocation: ChannelAllocation[];
  metadata: CampaignMetadata;
  generatedAt: string;
}