# lib/api-zod/src/generated/types/channelAllocation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChannelAllocationPriority } from "./channelAllocationPriority";

export interface ChannelAllocation {
  channel: string;
  percentage: number;
  estimatedMonthlySpend: string;
  priority: ChannelAllocationPriority;
}